将现实气泡的大小变为原来的样子

+++ src/shadowcasting.cpp
@@ -11,7 +11,7 @@
 #include "point.h"
 
 struct slope {
-    slope( int_least16_t rise, int_least16_t run ) {
+    slope(int_least8_t rise, int_least8_t run) {
         // Ensure run is always positive for the inequality operators
         this->run = std::abs( run );
         if( run < 0 ) {
@@ -23,8 +23,8 @@ struct slope {
 
     // We don't need more that 8 bits since the shadowcasting area is not that large,
     // currently the radius is 60.
-    int_least16_t rise;
-    int_least16_t run;
+    int_least8_t rise;
+    int_least8_t run;
 };
 
 static bool operator<( const slope &lhs, const slope &rhs )
